+++ src/Exceptions/DatabaseOperationException.cs
@@ -1,0 +1,5 @@
+using System;
+
+namespace PollingOutboxPublisher.Exceptions;
+
+public class DatabaseOperationException(string message, Exception exception) : Exception(message, exception);